requirements.txt
aiogram==3.4.1
aiohttp
asyncpg
apscheduler
python-dotenv